Рефактор функций для авторизации
- Разделена логика авторизации админов и пользователей
- Частично устранено дублирование кода

+++ bot/handlers/authorization/process_role.py
@@ -1,24 +1,27 @@
 import asyncio
-
+from aiogram.exceptions import TelegramBadRequest
 from aiogram.fsm.context import FSMContext
 from aiogram import F, Router
 from aiogram.types import Message
 
-from app.api.handlers.normalize import normalize_phone
-from bot.utils.exceptions import UserNotExistsException, IncorrectPhone
+from app.api.utils.auth import AuthUtils
+from app.db.base import async_session_factory
+from bot.utils.exceptions import UserNotExistsException
+from bot.utils.exceptions import IncorrectPhone
+from bot.utils.exceptions import IncorrectPasswordException
 from bot.keyboards.customer import CustomerKeyboards
 from bot.keyboards.backbuttons import BackButtons
-from bot.utils.fetch_user import fetch_user_by_phone
 from bot.utils.bot_utils import BotUtils
-from bot.states.admin import AdminAuth
+from bot.states.admin_auth import AdminAuth
+from bot.states.auth import Auth
 
 router = Router()
 
 
-@router.message(AdminAuth.waiting_for_phone)
+@router.message(Auth.waiting_for_phone)
 async def process_role(message: Message, state: FSMContext):
     """
-    Проверяет роль пользователя.
+    Проверяет роль пользователя в базе данных.
     
     Если роль админ, то запрашивает его пароль.
     """
@@ -28,28 +31,52 @@ async def process_role(message: Message, state: FSMContext):
     
     data = await state.get_data()
     last_bot_message_id = data.get("last_bot_message")
+    await BotUtils.delete_prev_messages(message, last_bot_message_id)
     
-    
-    try:
-        phone = await normalize_phone(message.text)
-        user = await fetch_user_by_phone(phone)     
-        await BotUtils.delete_prev_messages(message, message_id=last_bot_message_id)
-        await message.delete()
+    async with async_session_factory() as session:
+        try:
+            role, obj, phone_number = await AuthUtils.process_role_in_db(phone_number=message.text, session=session)
+            
+            if phone_number is None:
+                raise IncorrectPhone(IncorrectPhone.__doc__)
+            
+            if obj is None:
+                raise UserNotExistsException(UserNotExistsException.__doc__)
+            
+            
+            await BotUtils.delete_prev_messages(message, message_id=last_bot_message_id)
+            await message.delete()
+            
+        except (UserNotExistsException) as e:
+            sent = await message.answer(str(e), parse_mode="HTML")
+            await asyncio.sleep(2)
+            await message.delete()
+            await sent.delete()
+            return
+
+        except (IncorrectPhone, IncorrectPasswordException) as e:
+            sent = await message.answer(str(e), parse_mode="HTML")
+            await state.update_data(error_message=sent.message_id)
+            return 
         
-    except (IncorrectPhone, UserNotExistsException) as e:
-        sent = await message.answer(str(e), parse_mode="HTML")
-        await asyncio.sleep(2)
-        await message.delete()
-        await sent.delete()
-        return
+        await state.update_data(
+            phone=phone_number,
+            role=role,
+        )
     
     
-    await state.update_data(phone=phone)
+        data = await state.get_data()
+        error_message = data.get("error_message")
+        try:
+            if error_message:
+                await BotUtils.delete_prev_messages(message, error_message)      
+        except TelegramBadRequest:
+            pass
     
     
-    if user.role == "admin":
+    if role == "admin":
         sent = await message.answer("Пожалуйста, введите ваш пароль для подтверждения доступа.", reply_markup=await BackButtons.back_to_phone())
-        await state.set_state(AdminAuth.waiting_for_password)
+        await state.set_state(AdminAuth.password)
         await state.update_data(last_bot_message=sent.message_id)
         await asyncio.sleep(2)
     else:
